Add CI test using `dry-run` input
Tests a sample `stack.yaml` that resolves to the default LTS Haskell snapshot. This will mitigate the likelihood of regressions when updating this action to use a different snapshot repository.

The action is split up to avoid building with an updated stack.yaml, since this would require a more elaborate test case. However, committing the result is useful to test and doesn't require any additional scaffolding.