Add CI workflow (#1)
* Add CI test using `dry-run` input

Tests a sample `stack.yaml` that resolves to the default LTS Haskell snapshot. This will mitigate the likelihood of regressions when updating this action to use a different snapshot repository.

The action is split up to avoid building with an updated stack.yaml, since this would require a more elaborate test case. However, committing the result is useful to test and doesn't require any additional scaffolding.

* Restyled by prettier-yaml (#2)

Co-authored-by: Restyled.io <commits@restyled.io>

* Create valid test package

* Point build to stack-yaml input

* Consolidate steps and remove dry-run input

* Validate no build errors

* Remove deprecated set-output command

* Validate commit message

* Fix test variable

* Use delimiter for multiline output

A randomized delimiter is used for security reasons.

See: https://github.com/orgs/community/discussions/26288#discussioncomment-3876281

* Fix `message` output

* Test other action outputs

---------

Co-authored-by: restyled-io[bot] <32688539+restyled-io[bot]@users.noreply.github.com>
Co-authored-by: Restyled.io <commits@restyled.io>